+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* Incompatibilities in the ISO 19115 JSON Schema between JavaScript and Python validation libraries
+  [#243](https://gitlab.data.bas.ac.uk/uk-pdc/metadata-infrastructure/metadata-library/-/issues/243)
+
 ## 0.13.1 - 2024-10-02
 
 ### Fixed